Update patch set 2
Patch Set 2:

As far as the verification is concerned, same as Patch Set 1:

This merges and builds fine on the open-source master. However, it conflicts with Google's internal master tree where it appears that the same area of the code was also modified.

Reviewers will need to consider that conflict as part of their review, including whether it will be possible to merge this change at all and how to resolve the conflict.

Patch-set: 2