Update patch set 1
Patch Set 1:

The conflict will only be a name change. So can easily be resolved at compile time. 

If you would like us to handle it we can do it as well. Approve and merge one change first, we can rebase and apply the new change on top of the one that is merged.

Patch-set: 1