Update patch set 1
Patch Set 1:

Alex answered my question on why move/rename:

Reason 1: Using gsm package in CDMAPhone isn't right.
Reason 2: Cat is the base specification for both Stk and Ccat. At this
point all the Stk implementation we have is common for both Ccat and Stk
(because these requirements come from Cat spec). However, when we add
support for more commands - they will be specific to Stk or Ccat (and be
different between Stk and Ccat). So we would want to create derived
classes for Ccat and Stk based on top of Cat where we will implement these
differences. If we keep using Stk package for that all such implementation
will be hacks and will look really ugly (not something we want for
Android).

Thanks,
Alex

Patch-set: 1