Update patch set 2
Patch Set 2:

This looks reasonable to me.  I'd like to see hard coded values like this moved to resources so they can be customized for products via overlays.  Of course this patch will need to be changed to use the new LightsService class instead of HardwareService (which was removed in Froyo).

Patch-set: 2
Label: Code-Review=0
Label: Verified=0