Update patch set 2
Patch Set 2: Verified

The style looks reasonable, the change merges and builds fine on the open-source tree, merges fine into Google's internal tree.

I'm not quite sure what order to expect. On the full-eng emulator, I'm getting (in this order) the Chinese IME (rendered with Chinese characters, my apologies for not copying them here), the Japanese IME and the Android keyboard, in that order, and this might not be the expected behavior.

Patch-set: 2
Label: Verified=+1