Update patch set 2
Patch Set 2:

Hi,

Thanks for kenny and enf1234567890's comments.

The intention of this patch was the order of IMEs are inconsistent among different users even they installed the same set of IMEs. So user A may see "倉頡輸入法", "Android keyboard", "注音輸入法". User B may see "倉頡輸入法", "注音輸入法", "Android keyboard". (My apologies for typing Chinese here. I want to show that the user A's list looks odd for Chinese users. Those Chinese IMEs are not grouped together)

So the point is the orders are better to be consistent. Not what the order is. List those Chinese IMEs before or after the Android keyboard are both fine.

I tried to figure out what Collator is and I found it's a locale-sensitive string comparator. I think it is not necessary for the original intention since the String type is storing UTF-16 that is ordered by languages and glyphs already.

Please let me know if I still need to change to use Collator. Thanks.

Patch-set: 2
Reviewer: Gerrit User 1004123 <1004123@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0