Update patch set 3
Patch Set 3:

This merges and builds fine in the open-source tree.

Just as an FYI, there's a large number of conflicts with Google's internal master tree, such that merging this is will require some extra effort. I'd like this change to be verified by someone who'll be able to actually evaluate that merge conflict better and decide whether this'll be practical at all.

Patch-set: 3
Label: Code-Review=0
Label: Verified=0