Update patch set 1
Patch Set 1:

Are you talking about the comment relating to the margin?
The margin should be from the left since most app developers use the leading margin as a left margin to design the layout.

This works very well with the MMS app without needing to change it.
I believe the API should be updated to add a left and right margin classes to enable developers of BiDi aware application to define margins.

This patch set makes it so that bidi appears properly on every application that uses text layout, as long as there is more RTL than LTR text in the line.

Patch-set: 1