Update patch set 1
Patch Set 1: Do not submit

>The margin should be from the left since most app developers use the leading margin as a left margin to design the layout.

in RTL the leading margin is on the right. If an app is making the wrong assumption, the app has to be fixed. RTL layouts are currently not supported on Android and it's something we've started working on, precisely so that apps can support leading padding/margins, etc.

Patch-set: 1