Update patch set 1
Patch Set 1:

Considering you have changes that affect RTL, every change I make might be rejected since it is already fixed internally...

Is there a way I could get information about or access to the code changed internally? An NDA or something?

I read somewhere that one of the main focal points for gingerbread is UI revamp, so it makes sense these things changed, but I don't want to have to make workarounds in the next version like I do in FroYo and done in Eclair.
I'd like you to have perfect BiDi support in stock Gingerbread, and I believe I can add it in a month or so.

Patch-set: 1