Update patch set 1
Patch Set 1:

The change was originally made to address the following problem:

Sometimes phone doesn't perform power down registration on 1x. After looking at the logs we found that it was due to the data call release that happened. As per the CDMA state machine, after call release UE needs to go to Initialization state. If the RADIO_POWER(OFF) request was sent in this state before reaching Idle state, there is a chance that modem does not perform power down registration. 

That was probably the reason why there is a provision in the CDMA standard for the UE to indicate power down in the call release message itself. But since Android first releases the data call without indicating that it is due to powering down the phone, modem is unable to set the bit in the call release message. Instead if we directly power down, the call release will happen with the proper indication and the power down will be much faster as well.

Patch-set: 1