Update patch set 1
Patch Set 1:

In that additional cleanup needs to be done when a power down occurs sounds like a bug in the RIL. Also, the power down could happen at "anytime" so this situation could happen at not mater what we do.

So how big is the problem that's being solved, how often does it occur?

If this is a "big" problem, another possible solution would be to create a new contract, for instance create a new command with more explicit semantics. 

Other idea's anyone?

Patch-set: 1