Update patch set 1
Patch Set 1:

Phone needs to do power down registration so that the base station can release its resources and also know that incoming calls cannot be routed to this phone anymore. When it powers up again, it will do power-up registration. So if we miss out on doing power-down registration sometimes, it is a serious problem from carrier perspective.

Which part of the contract are we now referring to?
1) disconnect data call after RIL has indicated that the call is already torn down
2) sending power-down without disconnecting data call.

If it is the former, I think it is more of an optimization. Why ask RIL to tear down the call when it has indicated that it is already torn down?

Patch-set: 1