Create change
Patch-set: 1
Change-id: Iad71e78b1c72df795c83f831d1ff05290b5a1288
Subject: Store the event arguments, an array of Integer and String, as an Intent extra with no conversion to another type.  It contains Integer and String, so it can be serialized with no risk of throwing a runtime exception.
Branch: refs/heads/master
Commit: 69147650328b2568a68b6292159599c2799543e8