Update patch set 1
Patch Set 1:

Thanks for the patch. It is indeed a bug that needs to be fixed but I'd rather not fix it this way. The SelectionNotifier should not be a handler in the first place. I'll take a closer look at this tomorrow and see what a proper fix would be.

Patch-set: 1