Shutdown when capacity is 0% and non charging or when battery is dead
The Android framework does not shutdown when battery capacity is 0% and
a (non-charging) charger is attached (USB or AC). This handling is wrong
since a charger might very well be attached but charging has stopped
because USB suspended or the charging algorithm has stopped because of
battery safety handling. Also shutdown when battery is reported 'dead'.
This might happen although charging is present.

Change-Id: I6e685d0ae3c838c2ae66bb97c49b67ae4003b17e