Update patch set 1
Patch Set 1: No score

(1 inline comment)

Naveen,

Lets revert the change to CdmaDataConnectionTracker as I think we're in agreement its "just" an optimization. We can always look at it in the future. Your comment of the registration seems reasonable, but I'd like you to review the moving of dcTracker.sendMessage.

-- Wink

Patch-set: 1
Label: Code-Review=0