Update patch set 1
Patch Set 1:

I am not sure which part your question is about, but I will explain a little Further. The no-cache means that the thing can be cached, but must be revalidated On every request.

The code today, if it receives the header:

Cache-Control: no-cache, max-age=200000

Will set expires = 200000 in the parseHeaders() function. When expires is set To 200000 the content will not be revalidated during this period which the it Should be done according to the no-cache directive.

The fix sets expires to 0 which means that content in cache is expired, but That only means that it should be revalidated, not that we can not have it in the cache.

Compare o the no-store case in the parseHeader() method that returns null Not caching the item.

Patch-set: 1
Label: Code-Review=0