Update patch set 1
Patch Set 1:

Wink, not sure I follow:  If an app wants the name of the registered network, it should call getNetworkOperatorName().  If it wants the name of the operator that issued the SIM, it should call getSimOperatorName().  These two are often the same, but can be different if the phone is roaming.

Patch-set: 1