Fix for truncated texts in power dialog
The airplane mode status texts in the power dialog is truncated
for certain languages (such as Greek).
This is due to hardcoded size values for the dialog items.
The solution is to change to dynamic sizing of the dialog items.

We change from hardcoded values to wrap_content and add the
previous hardcoded size as a minHeight attribute instead.

Change-Id: I1d9437b7b10a4cf2e0a11dfaaf7267b88746364c