Default WiFi sleep policy
Creates a defult.xml setting for WiFi sleep policy, and sets
this to never.

The motivation behind this is to improve user experience of WiFi
as the phone will not swich back to 3G as much.

A downside is that it may have impact on battery life, for
low-power mode this has little consequence in practice though.

Any way the setting is now possible to configure.

Change-Id: Ie459b8e70fdbc7c605452fe0692d7bc26460e939