Update patch set 3
Patch Set 3:

Romain / Jean-Baptiste

Apologies for not getting back sooner on this issue.  ListView, as was suggested by Romain, does not support key functionality that the gallery has which we're trying to leverage in this patch.  This includes the ability to set unselected alphas for items that are not the selected item.  It also does not have the ability to center on the selected item.  This patch shouldn't be decreasing the efficiency of Gallery either, only providing extra functionality for vertical scrolling in the gallery so that we can utilize the two key features I've mentioned above.  We would just build on top of the gallery but we cannot because most of the functionality necessary to implement vertical scrolling lies as private functionality inside of Gallery.  

We can add comments to the patch stating to use a ListView for its efficiency benefits if the user is not looking for those two features described, but we believe this is an improvement that should be considered as it does provide benefits a ListView does not.  We don't mean for this patch to take ListViews place in any way, only providing functionality for user's that may not require ListViews benefits (efficiency) due to smaller lists.

Thoughts?

Thanks,

Roberto

Patch-set: 3