Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

This patch looks reasonable, but I find it surprising that a GSM carrier would provision a SIM with neither the CPHS record nor the 3GPP record to store the call forwarding flag. If that's the case, then this seems like a reasonable way to deal with it.

Patch-set: 1