Update patch set 1
Patch Set 1:

Thanks!

Like Wink said, please rename the class to EONS. We've sadly not done a great job at following this convention (see SIM and GSM), but we're still trying to follow it as much as possible for new code.

The copyright/license notices for the new files look fine (yay!), and the code merges and builds fine in the open-source tree, and merges into Google's internal tree with only a trivial conflict.

Patch-set: 1