Update patch set 3
Patch Set 3:

Hi Roberto,

We really don't want to throw feature after feature in to the framework widgets.  Every such added feature is yet more code that we have to live with forever, maintain forever, and generally makes it that much more difficult to evolve the platform in the future.

There is *nothing* special about a widget in the framework.  If the framework widget doesn't do something you want, feel free to build a customized version in to your own app that does what you want.

Besides which, in many ways Gallery sucks (such as its scrolling behavior), so adding features to try to get more people to use it is not such a positive thing.  Certainly not without first fixing the existing features it has to work better.

Regardless of the code, we do not want such a feature in the framework at this point.  You are welcome to fork the gallery widget into your own widget with all of the features you want, and share it with anyone you want.  That is really a better approach for everyone: it means you don't need to be tied to a platform version to get those features, you can easily improve and evolve the features without backwards compatibility issues, and the framework doesn't need to continue to maintain support for the features as originally implemented for the rest of its life.

Patch-set: 3