Update patch set 1
Patch Set 1:

From author:

"There is a PROXY_CHANGED_ACTION which we already listening to. If the idea is to also send out that broadcast whenever WiFi is connected, then yes that could work. However today this action is strictly used when the proxy SETTING is changed in the data base. Wouldn't it be risky to change the behaviour when this broadcast is used?
Another problem is that there is also a race condition involved, the connectivity manager also listens to WifiManager.NETWORK_STATE_CHANGED_ACTION so when we call setProxyConfig the connectivity manager still returns that a Mobile network is connected. Therefore we need to ensure that if we only get a PROXY_CHANGED_ACTION the connectivity manager is updated before us. Sounds a bit dangerous to me so therefore I thought listening to the WifiManager.NETWORK_STATE_CHANGED_ACTION would be the safest solution."

Patch-set: 1
Reviewer: Gerrit User 1002951 <1002951@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0