Update patch set 1
Patch Set 1:

The PROXY_CHANGE_ACTION message is being misused.  It is the only thing a proxy user should need to listen to.  It should be sent out when either the proxy value for the current network changes or when the current network changes.  This is a bug today.  Expanding the list of what everybody needs to listen for to figure proxy values is moving the wrong direction.

Yes there will always be a race - if somebody tries making a network connection when we're switching networks you will also have a race.  This is unavoidable.  If somebody tries to make a connection after the network is switched but before the PROXY_CHANGE_ACTION broadcast is processed they will get the wrong proxy and probably fail to connect.  If they had been a fraction earlier they would have been on the wrong network and been reset.  Network transitions are painful all around.

Patch-set: 1