Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

(1 inline comment)

Note my new convention is to -1 changes to that they get "marked" in the dash board. Otherwise it looks like I haven't reviewed it. It was a little startling when Robert Greenwalt started doing that to me. I'd suggested a change to Gerrit but Shawn, the Gerrit/Repo developer, and he said they use the same technique (-2 if not likely, -1 if likely but not yet, 0 if not reviewed and +1/+2 when ready).

So the -1 implies "not yet ready to submit".

Patch-set: 2
Label: Code-Review=-1
Label: Verified=0