Update patch set 3
Patch Set 3:

Hi Johan, "Clear data" does a "Force stop" as part of its operation, so all processes that could access this data are killed as part clearing it.  (It wouldn't be good to leave processes running as we rip the data out from under them.)

I'd really like to understand the situation causing this to be an issue.

Patch-set: 3
Reviewer: Gerrit User 1000559 <1000559@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0