Update patch set 3
Patch Set 3:

In continuation to the previous comment, the RIL_REQUET_OPERATOR can itself trigger EONS algorithm again because of a change in PLMN and this may not derive an EONS name for this PLMN. This will result in NO operator name since we already discarded RIL operator name in flavor of already existing EONS name which is now not valid.
There may be similar other cases, should we consider updating the ss.operatorAlphaLong() with EONS name?

Patch-set: 3