Discarding duplicate part of concatenated SMS message upon reception
If receiving the same part of a concatenated SMS message several times,
the duplicates were all stored. This could lead to a crash further on
since the message could be considered complete before all parts were
actually received, causing some PDUs to be null in the dispatched intent.
The change adds a check to see if the same part is already present, and
if so simply discards the duplicate.