Update patch set 1
Patch Set 1:

As Brad suggested in his inline comment, the SMS status constants are stored in a column in an SMS/MMS SQLite table, so unfortunately we can't just change the status constants to the correct values without also updating the SMS/MMS database with the new column values.

I've been thinking about a fix that would upgrade the SQLite database version number and modify the status column to the correct values. Since most of the messages in a typical database have status 0, it will only require a select and update for the messages with a non-zero status.

In the meantime, the error handling code mostly works correctly, even though the values are incorrect compared to the GSM spec, so changing them to the correct values without handling backwards compatibility of any preexisting SQLite SMS database would likely do more harm than good.

I'll come back to this bug in a week or two.

Patch-set: 1
Reviewer: Gerrit User 1003544 <1003544@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0