Fix for missing status reports for delayed messages
When an SMS is sent to a phone that is not turned on, the sending
phone receives a status report saying that the message is pending.
The pending intent that is to be matched with the status report
should then not be removed from the list but rather remain in the
list to await a success or fail status report.

Note that the status report notification will still be lost if the
sending phone is turned off before receiving the status report.