Update patch set 1
Patch Set 1:

Well, TP-ST status report values are indeed stored in the SQLite database (TelephonyProvider), but they are always stored as received in the SMS PDUs. To the best of my knowledge, the actual constants are never used for *storing* any values, only for comparing with received values in order to determine if a received value falls into the "delivered", "failed" or "pending" categories.

Looking at the packages/apps/Mms application (DeliveryReportActivity) and comparing it to the table in 3G TS 23.040, it is evident that the constants are used as the lower boundary of each category. This means that only values 128-255 are considered "failed" while 64-127 are considered "pending", while in fact failed values should be 64-255 and pending 32-63. The most significant bit of TP-ST is reserved for future use so values greater than or equal to 128 will probably never even be received, but the spec says that "The MS shall interpret any reserved values as 'Service Rejected' (01100011)" so it’s OK to consider any value greater than or equal to 64 as failed.

The end result is that if an SMS fails to be delivered, it will never be marked as failed in the MMS app, instead the status will be marked as pending indefinitely. Users will not know when their messages have failed permanently. This renders the delivery report feature half-useless.

I think it would be safe to change these constants as it would not impact what’s stored in any database, only what the behavior will be upon reception of a status report.

Patch-set: 1