Stagefright: Removed memcpy in playback datapath
During code review, it was found that the overlay buffers were not shared with the video decoder and hence
the decoded output was copied to the overlay buffers for rendering. Changes have been made to the framework
to remove this memcpy. The overlay buffers are shared with the video decoder.

Added the API getBuffers() to the following classes:

BpOMXRenderer
IOMXRenderer
OMXRenderer
VideoRenderer
SharedVideoRenderer
SoftwareRenderer
AwesomeRenderer
AwesomeRemoteRenderer
AwesomeLocalRenderer

Added the API setBuffers() to the following classes:

MediaSource
OMXCodec

Changes made to AwesomePlayer:

1. Modified the sequence in which the various components were initialized.
Now, the renderer is initialized before the codec is initialized.
2. The overlay buffers are shared with the video decoder.

Changes made to OMXCodec:
Modified the component quirks. Use OMX_UseBuffer(overlay buffers) for video decoder for output port.

Change-Id: I94a807f73deebfb1153f74a66c0dc98f9960f63a
Signed-off-by: Anu Sundararajan <sanuradha@ti.com>