Update patch set 1
Patch Set 1:

This commit looks very good and is a great improvement for ME-depersonalization. 
 
There is however a problem with this commit since it does not support the possibility to make emergency calls when the device is network locked. 
This is because the dismiss button is removed.
Also, the depersonalization UI will be shown for a network locked phone with wrong SIM card, but what if the SIM is absent?

Patch-set: 1
Reviewer: Gerrit User 1004876 <1004876@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0