Update patch set 1
Patch Set 1:

I'd rather we return an empty list in these cases.  It's less surprising for callers and simplifies their code.

Then we can fix the docs to say it never returns null.

But I suppose people need to know that it can return null on old/existing SDK versions, so.... maybe this is best.  Or both?  Keep docs saying it may return null, but try not to going forward?

Patch-set: 1