Update patch set 1
Patch Set 1:

Given the confusion evident in the comments here, I'd think a comment would be useful in the code near the "&& !noCache" to explain the subtleties.

RFC 2616 says "If a request includes the no-cache directive, it SHOULD NOT include min-fresh, max-stale, or max-age.", so I imagine the same goes for responses, and ignoring max-age in this case is okay.

Patch-set: 1
Label: Verified=0