Update patch set 3
Patch Set 3:

current.xml would definitely have to be updated in any case; at the point of submission I wasn't aware of that.  Will be included in the next patch set.

As for API changes. I would have thought simply adding a return type to a method (rather than changing an existing return type or removing a return type) would not break the API.  I.e., adding a return value should be backwards compatible.  If however you think it is an issue I would suggest adding a method startServiceWithReturnValue (better name anyone?) that does return the service's return value.

As for the additional flags parameter.  I think ServiceTestCase should support starting a service with specific start flags.  Otherwise there is no way for a unit test to verify a service's behavior on restarts and intent retries.  Are you fine with me leaving the additional parameter in there?

If this change passes (after the return value issue has been resolved), is there way for you to merge it with Google's internal version?

Patch-set: 3