Update patch set 1
Patch Set 1:

From Gloria: If mAudioPlayer->start does not return OK, is it still ok to set the playstatus as play? 
Probably we need to set the status as stop before "return err"?

From Aimi: Your guess is completely true.
We need to set the status as “STOP” before “return err”.

Patch-set: 1