Update patch set 1
Patch Set 1: Verified

The style looks good, and the change merges and builds fine in the open-source tree.

In Google's internal master tree, the underlying file was entirely deleted. I'll let the reviewers comment on whether it'd make sense to accept that change without merging it internally, or to accept it and replicate its logic wherever that code now lives in our internal tree.

Patch-set: 1