Update patch set 2
Patch Set 2: Fails; Do not submit

Returning a valid application info is certainly valid and needed. We've actually done so internally, as we are doing major changed to the layoutlib implementation. We do intend to backport this new layoutlib to older 2.x versions later one (early next year).

I'm not sure why it is needed to set targetSdkVersion on it. We haven't had to. In any case, from the other CL, the targetSdkVersion is not the compilation target, but the content of the targetSdkVersion attribute in the manifest.

Patch-set: 2
Reviewer: Gerrit User 1001414 <1001414@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2
Label: Verified=-1