Update patch set 2
Patch Set 2:

> Returning a valid application info is certainly valid and needed. We've actually done so internally, as we are doing major changed to the layoutlib implementation

Where ? BridgeContext.getApplicationInfo() always return null

> I'm not sure why it is needed to set targetSdkVersion on it. We haven't had to. 

The problem (NPE) happens in the TabWidget.initTabWidget method. See line 114
...
 if (context.getApplicationInfo().targetSdkVersion <= Build.VERSION_CODES.DONUT) {
...
When Eclipse opens layout file with TabWidget, it creates TabWidget and call initTabWidget (from constructor). In that moment context.getApplicationInfo() is null and opening the layout file fails (context is instance of BridgeLayout).

You can test this if you set breakpoint in the initTabWidget and open a loyout file with TabWidget.

> In any case, from the other CL, the targetSdkVersion is not the compilation target, but the content of the targetSdkVersion attribute in the manifest.

This solution is only when GraphicalEditorPart calls computeLayout (projectKey is Integer and the applicationInfo is created in BridgeContext's constructor.) For other case, projectKey can be other object and this code will be skipped. 

Anyway, would I abandon this changes ?

Patch-set: 2