Update patch set 1
Patch Set 1:

On a Nexus One, I get 18 seconds of the static boot logo, then 14 seconds of boot animation, and then the home screen.
On another phone it's 12 seconds of static boot logo and 6 seconds of animation.
So this boot sound would be played kind of late in the boot process, and not for very long. Would it be better to embed the sound in the boot ROM and talking directly to the driver?
That would allow the sound to be played much earlier.

Patch-set: 1