Update patch set 2
Patch Set 2:

We don't experience such delays on our hardware and have been discussing this internally today. Maybe mp3 in your case is decoded by a dsp and the delay is caused by loading the dsp firmware? I added support for playing .ogg-files also as these should play faster.
Also it is important that the file is fairly small and same length in time as the animation. In our case the sound is 2.5 seconds long and 68kb in size.

Patch-set: 2