Update patch set 2
Patch Set 2:

The delay I was talking about is the delay between power-on and when the boot animation runs, without your change.
The boot animation is started after the system_server runs (because the boot animation needs surfaceflinger, and surfaceflinger runs in the system server). On the devices I tried, it takes 12-18 seconds from the time I press the power button to the time I see the boot animation, and so with your change, it would take 12-18 seconds from the time I press the power button to when I would hear the boot sound, after which it would play for 6-14 seconds before the boot animation is done and the home screen appears.

Note that I don't have anything in particular against this change, I'm just wondering if there's a better way to do this that would allow the sound (and possibly the animation itself) to play much sooner after power-on. It seems some manufacturers are already doing this, since they play animations with sound almost as soon as the screen turns on.

Patch-set: 2