Update patch set 2
Patch Set 2:

You are right that it would be possible to do things earlier in the boot sequence. In this case we want the sound to be in sync with the animation, so playing the sound earlier would also mean doing the animation earlier. Such a change would be far more intrusive than the change in bootanimation. Possible, yes, but a different use case.

Patch-set: 2