Corrected scroll behavior for right-to-left paragraphs.
Paragraphs with RTL direction was misplaced when pressing
and dragging on the text field. The alignment used in Layout have no
meaning without the paragraph direction. ALIGN_NORMAL for a Latin text
is left aligned and ALIGN_OPPOSITE is right aligned but ALIGN_NORMAL
for Arabic is right alignment and ALIGN_OPPOSITE is left aligned.

The scrollTo method in Touch.java made the assumption that ALIGN_NORMAL
always is left alignment and ALIGN_OPPOSITE is right which is wrong.

Change-Id: I5033dac1f10b86365d1bce6afec6a7b84b05dcf9