Handle BrowserFrame is null in WebViewCore's contentDraw
WebViewCore's method contentDraw doesn't handle the case
when BrowserFrame (mBrowserFrame) is null. If BrowserFrame
is null and method contentDraw is executed a NullPointerException
will be thrown causing the appplicationto crash.

This patch checks BrowserFrame against null before it's used
in order to avoid a crash.

The problem has not been able to reproduce but it has happened
since found by a live user. The call stack indicates that a timer
has fired in JWebCoreJavaBridge, which has triggered a contentDraw.
In the meantime WebViewCore has been destroyed, which sets
WebViewCore's BrowserFrame to null.

Change-Id: I9aa6262f19993be0f0d79eb0dd0fbfc96d30f3ce