orientationchange event sent too early
The orientationchange event is sent to native webkit
before the screen dimensions have been reported to native webkit.
Thus, scripts using the orientationchange event will
use wrong dimensions.

This patch sends the orientationchange event to native
webkit after the new dimensions have been reported.

Reproduce this problem by create a website which
uses the orientationchange event. Read the window.innerWidth
and window.innerHeight and print to console in the
orientationchange eventhandler. Load the website into the
browser rotate the screen. Without the patch the
width and height values are wrong, with the patch
widht and height are the correct.

Change-Id: Id6e3f1f50be8f811942d120d955dfde9ad43662b