NPE fix in StatusBarService.makeExpandedVisible
The crash occurs when you trying to expand the statusbar
and the TrackingView hasn't been attached to window yet.
The fix is to only allow this when the TrackingView
has been attached to window.

You can reproduce this easy by trying to expand
the status bar during boot
(try to expand the status bar as it was visible).

Change-Id: I20f4069c6a03bff7dde9d6d98c56fdcd6c36426a