Fix issue #13366. Take account of stackFromBottom in computeVerticalScrollOffset().
Now on master branch.

I have tested it before on the froyo branch in the emulator via eclipse, however often I cannot get the master branch to build in eclipse, or before that in the terminal via make even. Any advice? I presume it is just the state of the master branch that it does not always compile. This resubmit to the master branch has not been tested an sich, but since the code is identical and the change is localized, the risk is minimal.

Change-Id: I984180af4b5e6a4a17714a662b43ffffc7d72fc2