Re-enabled initialization of sGpsNiInterface
The GpsNiInterface needs to be initialized whenever the GPS is
enabled. Without it, there will be no callback interface from the GPS
HAL to the JNI layer. The callback interface is needed for Control
Plane positioning request and for Network Initiated SUPL request on
platforms where the modem handles those messages.

Change-Id: I10fc48c0d581fe8771e31c820fd6233d9b66c056