Update patch set 2
Patch Set 2:

It was a deliberate decision at the time because the behavior of soft references was not what I wanted -- they would allow the heap size to grow, which we very much did not want with the memory constraints we had.

Before approving this I need to check whether either the behavior of soft references has changed since then, or we no longer are as concerned with memory use.

Patch-set: 2
Reviewer: Gerrit User 1000559 <1000559@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0