Update patch set 1
Patch Set 1:

Thanks Jean-Baptiste and Jaikumar for the prompt review and for all attention and details provided.

Since I don't have access to Android's internal master, could you please just confirm if the socket used for SCO is able to  stopped in case the requester of this socket wants to release it and open it again?

Patch-set: 1
Reviewer: Gerrit User 1005295 <1005295@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0